tests/security/PenetrationTests.ZAP/requirements.txt
requests>=2.31.0
python-owasp-zap-v2.4>=0.0.21
pyyaml>=6.0.1
pathlib2>=2.3.7
